refactor: split sqllogictest executor.rs into focused modules
Split the monolithic 991-line executor.rs file into 5 focused modules
for better maintainability and clarity:

- executor/core.rs (367 lines): Runner struct, configuration, and main execution flow
- executor/validator.rs (249 lines): Result validation and comparison logic
- executor/record_processor.rs (323 lines): Individual record execution
- executor/parallel.rs (95 lines): Parallel test execution
- executor/mod.rs (11 lines): Module declarations and public API

Benefits:
- Single Responsibility: Each module handles one aspect of execution
- Improved testability: Smaller modules are easier to test in isolation
- Better readability: Clear separation of concerns
- Easier maintenance: Changes to one aspect don't affect unrelated code

Public API remains unchanged (backward compatible).

Closes #1356
