+++ crates/sqllogictest/src/executor/mod.rs
@@ -1,0 +1,11 @@
+//! Core execution logic for sqllogictest runner.
+
+mod core;
+mod parallel;
+mod record_processor;
+mod validator;
+
+// Re-export public types and traits
+pub use core::{
+    AsyncDB, DB, Partitioner, Runner, RunnerLocals, default_partitioner,
+};
